+++ src/test/java/net/webpdf/wsclient/session/SessionTest.java
@@ -1,6 +1,6 @@
 package net.webpdf.wsclient.session;
 
-import net.webpdf.wsclient.session.access.UserAccess;
+import net.webpdf.wsclient.session.auth.UserAuthProvider;
 import net.webpdf.wsclient.webservice.WebServiceProtocol;
 import net.webpdf.wsclient.session.soap.SoapWebServiceSession;
 import org.junit.jupiter.api.Test;
@@ -25,7 +25,7 @@ public class SessionTest {
             URL url = new URL(SOME_URL);
             try (SoapWebServiceSession soapSession =
                          SessionFactory.createInstance(WebServiceProtocol.SOAP, url,
-                                 new UserAccess("usr", "pwd"))) {
+                                 new UserAuthProvider("usr", "pwd"))) {
                 assertNotNull(soapSession,
                         "SOAPSession should have been initialized.");
                 assertTrue(soapSession.isUseLocalWsdl(),